refactor:lib and configuration
