+++ Cargo.toml
@@ -20,19 +20,8 @@ name = "devapi"
 actix-http = "=3.0.0-beta.10"
 actix-web = "=4.0.0-beta.9"
 # Using table-like toml syntax to avoid a super-long line!
-[dependencies.sqlx]
-version = "0.5.7"
-default-features = false
-features = [
-"runtime-actix-rustls",
-"macros",
-"postgres",
-"uuid",
-"chrono",
-"migrate"
-]
 
-serde = "1.0.115"
+serde = { version = "1.0.115", features = ["derive"] }
 config = { version = "0.10.1", default-features = false, features = ["yaml"] }
 sqlx = { version = "0.5.1", default-features = false, features = [ "runtime-actix-rustls", "macros", "postgres", "uuid", "chrono", "migrate", "offline"] }
 uuid = { version = "0.8.1", features = ["v4"] }
@@ -53,6 +42,7 @@ tracing-actix-web = "0.4.0-beta.12"
 anyhow = "1.0.40"
 base64 = "0.13.0"
 argon2 = { version = "0.3", features = ["std"] }
+serde_json = "1.0.61"
 
 [dev-dependencies]
 once_cell = "1.7.2"
@@ -61,7 +51,6 @@ quickcheck = "0.9.2"
 quickcheck_macros = "0.9.1"
 fake = "~2.3.0"
 wiremock = "0.5"
-serde_json = "1.0.61"
 actix-rt = "2"
 tokio = { version = "1", features = ["macros"] }
 linkify = "0.5.0"
