+++ src/main.rs
@@ -1,4 +1,4 @@
-use devapi::run;
+use devapi::startup::run;
 use std::net::TcpListener;
 
 #[actix_web::main]
