+++ src/routes/mod.rs
@@ -1,0 +1,5 @@
+mod health_check;
+mod subscription;
+
+pub use health_check::*;
+pub use subscription::*;
